requirements.txt
psycopg2-binary
pandas
numpy
requests
sqlalchemy
yfinance
pydantic
fastapi
uvicorn
python-dotenv
pytest
boto3

# Airflow core and providers (example versions)
apache-airflow
apache-airflow-providers-postgres
apache-airflow-providers-http
apache-airflow-providers-sqlite
apache-airflow-providers-docker
apache-airflow-providers-ssh
apache-airflow-providers-google